Finish initial refactoring of Parser functions
